I'm not entirely happy with these changes.
The only ones that make sense are
1. Fix to call to log10 in VUMeter that tripped up newer GCC compilers.
2. Fix to RadioListener.cpp - the EXTERNAL reference switch was wired backward.
3. Fixes to documentation in SoDaRadio.md.

The RadioControl and USRPCtrl changes will be backed out in the next commit.

+++ gui_common/VUMeter.cpp
@@ -66,7 +66,7 @@ void VUMeter::updateLevel(float* buf, qint64 len) {
         sum_sq += s * s;
     }
     double rms = std::sqrt(sum_sq / (double)len);
-    float db = (rms > 1e-10) ? 20.0f * std::log10f((float)rms) : MIN_DB - 10.0f;
+    float db = (rms > 1e-10) ? 20.0f * std::log10((float)rms) : MIN_DB - 10.0f;
     target_db = db;  // timer applies ballistics, so just store the raw level
 }
